Removed test_system_registration.py
1. Tests from test_system_registration.py are moved to test_contenthost.py
2. test_system_registration is removed
